+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 ## 1.5.0
 
+* Vertical scroll with keyboard keys: arrow up, arrow down, page down and page up.
 * Added `visibleRowIndex` in `EasyTableCellBuilder`.
 * Added `notifyUpdate` method in `EasyTableModel`.
 
 ## 1.4.0
 
-* Restricted to Flutter 3.0.0 or higher.
 * Feature to display horizontal scrollbar only when needed.
   * A warning is being displayed in the console due to a bug in Flutter: https://github.com/flutter/flutter/issues/103939
   * The error happens when the horizontal scrollbar is hidden after being visible.
@@ -21,6 +21,7 @@
 
 ## 1.3.0
 
+* Updated to Flutter 3.0.0 or higher.
 * Improved renderer performance.
 * `CellThemeData.contentHeight` moved to `EasyTable.cellContentHeight`.
 * `EasyTableColumn`